Code cleanup LearningResourceFormAccordion
